IrLed class RedLed class shotControl Class af
