+++ main.cpp
@@ -1,28 +1,20 @@
 #include "hwlib.hpp"
 #include "rtos.hpp"
+#include "sendControl.hpp"
 
-int main(void) {
-    // IR output LED
-    // red output LED
-    auto red = hwlib::target::pin_out(hwlib::target::pins::d7);
 
-    // switch which enables the 36 kHz output
-    auto sw = hwlib::target::pin_in(hwlib::target::pins::d43);
 
-    // when the switch is pressed,
-    // repeat sending 1 ms signal and 1 ms silence
-    // and show this on the LED
-    for (;;) {
-        hwlib::wait_ms(1);
+extern unsigned char bmptk_stack[ 81920 ];
 
-        sw.refresh();
-        ir.write(!sw.read());
-        red.write(!sw.read());
-        ir.flush();
-        red.flush();
+extern unsigned int __stack_start;
+extern unsigned int __stack_end;
 
-        hwlib::wait_ms(1);
-        ir.write(0);
-        ir.flush();
-  }
+
+int main( void ){	  
+   
+   hwlib::wait_ms(1000);
+
+   auto sendTest = sendControl("sendTest");
+   sendTest.sendMesasge(0b1111111111111111);
+   rtos::run();
 }
